+++ app/models/GatewayAccount.class.js
@@ -1,12 +1,4 @@
-const GatewayAccountCredential = require('@models/GatewayAccountCredential.class')
-
-const CREDENTIAL_STATE = {
-  CREATED: 'CREATED',
-  ENTERED: 'ENTERED',
-  VERIFIED: 'VERIFIED_WITH_LIVE_PAYMENT',
-  ACTIVE: 'ACTIVE',
-  RETIRED: 'RETIRED'
-}
+const { GatewayAccountCredential, CREDENTIAL_STATE } = require('@models/GatewayAccountCredential.class')
 
 /**
  * @class GatewayAccount
@@ -18,6 +10,7 @@ const CREDENTIAL_STATE = {
  * @property {string} analyticsId - Google analyticsId of the gateway account
  * @property {boolean} toggle3ds - whether 3DS is enabled or not on this gateway account
  * @property {[GatewayAccountCredential]} gatewayAccountCredentials - available credentials for gateway account
+ * @property {GatewayAccountCredential} activeCredential - the active credential for the gateway account
  * @property {Object} rawResponse - raw 'gateway account' object
  */
 class GatewayAccount {
@@ -53,7 +46,8 @@ class GatewayAccount {
       this.gatewayAccountCredentials = gatewayAccountData?.gateway_account_credentials
         .map(credentialData => new GatewayAccountCredential(credentialData))
 
-      this.activeCredential = this.gatewayAccountCredentials.filter((credential) => credential.state === CREDENTIAL_STATE.ACTIVE)[0] || null
+      this.activeCredential = this.gatewayAccountCredentials.filter((credential) =>
+        credential.state === CREDENTIAL_STATE.ACTIVE)[0] || null
     }
     /** @deprecated this is a temporary compatability fix! If you find yourself using this for new code
      * you should instead add any rawResponse data as part of the constructor */
